Geometry classes (Draft)

+++ src/main/java/org/achrafsoltani/games/lzladx/engine/geometry/Position.java
@@ -2,28 +2,28 @@ package org.achrafsoltani.games.lzladx.engine.geometry;
 
 public class Position {
 
-    private float x;
+    private double x;
 
-    private float y;
+    private double y;
 
-    Position(float x, float y){
+    Position(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
